refactor(tests): Move setTaskStatus tests to dedicated file
