+++ tests/unit/task-manager.test.js
@@ -167,89 +167,6 @@ const testParsePRD = async (prdPath, outputPath, numTasks) => {
 	}
 };
 
-// Create a simplified version of setTaskStatus for testing
-const testSetTaskStatus = (tasksData, taskIdInput, newStatus) => {
-	// Handle multiple task IDs (comma-separated)
-	const taskIds = taskIdInput.split(',').map((id) => id.trim());
-	const updatedTasks = [];
-
-	// Update each task
-	for (const id of taskIds) {
-		testUpdateSingleTaskStatus(tasksData, id, newStatus);
-		updatedTasks.push(id);
-	}
-
-	return tasksData;
-};
-
-// Simplified version of updateSingleTaskStatus for testing
-const testUpdateSingleTaskStatus = (tasksData, taskIdInput, newStatus) => {
-	// Check if it's a subtask (e.g., "1.2")
-	if (taskIdInput.includes('.')) {
-		const [parentId, subtaskId] = taskIdInput
-			.split('.')
-			.map((id) => parseInt(id, 10));
-
-		// Find the parent task
-		const parentTask = tasksData.tasks.find((t) => t.id === parentId);
-		if (!parentTask) {
-			throw new Error(`Parent task ${parentId} not found`);
-		}
-
-		// Find the subtask
-		if (!parentTask.subtasks) {
-			throw new Error(`Parent task ${parentId} has no subtasks`);
-		}
-
-		const subtask = parentTask.subtasks.find((st) => st.id === subtaskId);
-		if (!subtask) {
-			throw new Error(
-				`Subtask ${subtaskId} not found in parent task ${parentId}`
-			);
-		}
-
-		// Update the subtask status
-		subtask.status = newStatus;
-
-		// Check if all subtasks are done (if setting to 'done')
-		if (
-			newStatus.toLowerCase() === 'done' ||
-			newStatus.toLowerCase() === 'completed'
-		) {
-			const allSubtasksDone = parentTask.subtasks.every(
-				(st) => st.status === 'done' || st.status === 'completed'
-			);
-
-			// For testing, we don't need to output suggestions
-		}
-	} else {
-		// Handle regular task
-		const taskId = parseInt(taskIdInput, 10);
-		const task = tasksData.tasks.find((t) => t.id === taskId);
-
-		if (!task) {
-			throw new Error(`Task ${taskId} not found`);
-		}
-
-		// Update the task status
-		task.status = newStatus;
-
-		// If marking as done, also mark all subtasks as done
-		if (
-			(newStatus.toLowerCase() === 'done' ||
-				newStatus.toLowerCase() === 'completed') &&
-			task.subtasks &&
-			task.subtasks.length > 0
-		) {
-			task.subtasks.forEach((subtask) => {
-				subtask.status = newStatus;
-			});
-		}
-	}
-
-	return true;
-};
-
 // Create a simplified version of listTasks for testing
 const testListTasks = (tasksData, statusFilter, withSubtasks = false) => {
 	// Filter tasks by status if specified
@@ -322,120 +239,6 @@ const { findNextTask, generateTaskFiles, clearSubtasks, updateTaskById } =
 describe('Task Manager Module', () => {
 	beforeEach(() => {
 		jest.clearAllMocks();
-	});
-
-	describe('setTaskStatus function', () => {
-		test('should update task status in tasks.json', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Act
-			const updatedData = testSetTaskStatus(testTasksData, '2', 'done');
-
-			// Assert
-			expect(updatedData.tasks[1].id).toBe(2);
-			expect(updatedData.tasks[1].status).toBe('done');
-		});
-
-		test('should update subtask status when using dot notation', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Act
-			const updatedData = testSetTaskStatus(testTasksData, '3.1', 'done');
-
-			// Assert
-			const subtaskParent = updatedData.tasks.find((t) => t.id === 3);
-			expect(subtaskParent).toBeDefined();
-			expect(subtaskParent.subtasks[0].status).toBe('done');
-		});
-
-		test('should update multiple tasks when given comma-separated IDs', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Act
-			const updatedData = testSetTaskStatus(testTasksData, '1,2', 'pending');
-
-			// Assert
-			expect(updatedData.tasks[0].status).toBe('pending');
-			expect(updatedData.tasks[1].status).toBe('pending');
-		});
-
-		test('should automatically mark subtasks as done when parent is marked done', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Act
-			const updatedData = testSetTaskStatus(testTasksData, '3', 'done');
-
-			// Assert
-			const parentTask = updatedData.tasks.find((t) => t.id === 3);
-			expect(parentTask.status).toBe('done');
-			expect(parentTask.subtasks[0].status).toBe('done');
-			expect(parentTask.subtasks[1].status).toBe('done');
-		});
-
-		test('should throw error for non-existent task ID', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Assert
-			expect(() => testSetTaskStatus(testTasksData, '99', 'done')).toThrow(
-				'Task 99 not found'
-			);
-		});
-	});
-
-	describe('updateSingleTaskStatus function', () => {
-		test('should update regular task status', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Act
-			const result = testUpdateSingleTaskStatus(testTasksData, '2', 'done');
-
-			// Assert
-			expect(result).toBe(true);
-			expect(testTasksData.tasks[1].status).toBe('done');
-		});
-
-		test('should update subtask status', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Act
-			const result = testUpdateSingleTaskStatus(testTasksData, '3.1', 'done');
-
-			// Assert
-			expect(result).toBe(true);
-			expect(testTasksData.tasks[2].subtasks[0].status).toBe('done');
-		});
-
-		test('should handle parent tasks without subtasks', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Remove subtasks from task 3
-			const taskWithoutSubtasks = { ...testTasksData.tasks[2] };
-			delete taskWithoutSubtasks.subtasks;
-			testTasksData.tasks[2] = taskWithoutSubtasks;
-
-			// Assert
-			expect(() =>
-				testUpdateSingleTaskStatus(testTasksData, '3.1', 'done')
-			).toThrow('has no subtasks');
-		});
-
-		test('should handle non-existent subtask ID', async () => {
-			// Arrange
-			const testTasksData = JSON.parse(JSON.stringify(sampleTasks));
-
-			// Assert
-			expect(() =>
-				testUpdateSingleTaskStatus(testTasksData, '3.99', 'done')
-			).toThrow('Subtask 99 not found');
-		});
 	});
 
 	describe('listTasks function', () => {
